Removed some scaling code in PLAGE calculation
There was some scaling code in the PLAGE calculation, specifically
dividing each pathway data by the pathway length and scaling the
first meta gene by the first eigenvalue. The idea here was to obtain
an output activity level that could be interpreted directly, perhaps
as an average of the metabolites on the pathway. This doesn't
work, so removing. The p-values obtained are identical.